+++ modules/system/settings.js
@@ -76,7 +76,7 @@ const registerSystemSettings = () => {
     scope: "client",
     config: true,
     type: Boolean,
-    default: true,
+    default: false,
     onChange: (value) => {
       LOGGER.log(`Changed automaticallyResizeSheets to ${value}`);
     },
@@ -112,6 +112,18 @@ const registerSystemSettings = () => {
     default: false,
     onChange: (value) => {
       LOGGER.log(`Changed debugLogs to ${value}`);
+    },
+  });
+
+  game.settings.register("cyberpunk-red-core", "debugElements", {
+    name: "CPR.debugelements",
+    hint: "CPR.debugelementshint",
+    scope: "client",
+    config: true,
+    type: Boolean,
+    default: false,
+    onChange: (value) => {
+      LOGGER.log(`Changed debugElements to ${value}`);
     },
   });