+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # DEV WIP CHANGES
-
+- added Debug Elements setting for developers
+- removed hp, humanity, and emp calculations for mooks
 
 # DEV WIP BUG FIXES
+- fixed #189
 
 # Version: 0.74.1 (Hot Fix) |  Date:  2021-04-25
 **Bug Fix**